requirements.txt
tensorflow-gpu==1.13.1
pandas>=1.0.5
numpy>=1.16.4
numba>=0.45.1
scipy>=1.3.1
deepctr[gpu]==0.8.6
catboost==0.26
lightgbm==3.2.1
gensim
tqdm